fix: Split bech32AddressHelpers
